refactor: move lexer and its dependencies into separate crate
