+++ compiler-lib/Cargo.toml
@@ -12,8 +12,13 @@ edition = "2018"
 license = "MIT OR Apache-2.0"
 
 [dependencies]
+diagnostics = { path = "../diagnostics" }
+compiler-shared = { path = "../compiler-shared" }
 asciifile = { path = "../asciifile" }
 symtab = { path = "../symtab" }
+strtab = { path = "../strtab" }
+lexer = { path = "../lexer" }
+utils = { path = "../utils" }
 failure = "0.1.3"
 termcolor = "1.0.4"
 derive_more = "0.13"
