+++ utils/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "utils"
+version = "0.1.0"
+authors = [
+    "flip1995 <>",
+    "hediet <>",
+    "joshuabach <>",
+    "problame <>",
+    "reiner-dolp <>"
+]
+edition = "2018"
+license = "MIT OR Apache-2.0"
+
+[dependencies]
+failure = "0.1.3"
+termcolor = "1.0.4"
+asciifile = { path = "../asciifile" }
